+++ src/components/Tab/Tab.tsx
@@ -1,10 +1,18 @@
 import { FC } from 'react'
+import { Icon } from 'react-feather'
 import { NavLink } from 'react-router-dom'
 
 import { Typography } from '@/components'
 
-import { createIcon, getClassName } from './core/tab.helpers'
-import { IProps } from './core/tab.typings'
+import { createIcon, getClassName } from './internal/Tab.helpers'
+
+export interface IProps {
+  className?: string
+  text: string
+  disabled?: boolean
+  iconStart?: Icon
+  link: string
+}
 
 export const Tab: FC<IProps> = ({
   text,
@@ -14,8 +22,8 @@ export const Tab: FC<IProps> = ({
   className,
 }) => {
   return (
-    <NavLink to={link} className={getClassName({ disabled, className })}>
-      {iconStart && createIcon({ icon: iconStart })}
+    <NavLink to={link} className={getClassName(disabled, className)}>
+      {iconStart && createIcon(iconStart)}
       <Typography element="span" font="poppins" weight="medium">
         {text}
       </Typography>
