--- src/components/Tab/core/tab.typings.tsx
@@ -1,9 +1,0 @@
-import { Icon } from 'react-feather'
-
-export interface IProps {
-  className?: string
-  text: string
-  disabled?: boolean
-  iconStart?: Icon
-  link: string
-}
