+++ src/components/Tab/index.ts
@@ -1,0 +1,1 @@
+export { Tab } from './Tab'
