+++ src/components/Tab/internal/Tab.helpers.ts
@@ -3,13 +3,11 @@ import { Icon } from 'react-feather'
 import { NavLinkProps } from 'react-router-dom'
 
 import s from './tab.module.scss'
-import { IProps } from './tab.typings'
 
-export const getClassName = ({
-  disabled,
-  className,
-}: Pick<IProps, 'className'> &
-  Pick<Required<IProps>, 'disabled'>): NavLinkProps['className'] => {
+export const getClassName = (
+  disabled: boolean,
+  className: string | undefined
+): NavLinkProps['className'] => {
   return ({ isActive }) => {
     const classNames: string[] = [s['tab'], 'f', 'f-ai-center', 'pb-2']
 
@@ -29,7 +27,7 @@ export const getClassName = ({
   }
 }
 
-export const createIcon = ({ icon }: { icon: Icon }): ReactNode => {
+export const createIcon = (icon: Icon): ReactNode => {
   return createElement(icon, {
     className: 'mr-1',
     size: 24,
