+++ src/components/Tab/internal/tab.module.scss
